+++ src/mintSDL/maths/maths.h
@@ -1,8 +1,8 @@
 #ifndef MATHS_H
 #define MATHS_H
 
-double mint_MathMin(double a, double b);
-double mint_MathMax(double a, double b);
-double mint_MathClamp(double a, double min, double max);
+float mint_MathMin(float a, float b);
+float mint_MathMax(float a, float b);
+float mint_MathClamp(float a, float min, float max);
 
 #endif
